% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\docType{class}
\name{BasicFilter-class}
\alias{AccnumFilter}
\alias{AliasFilter}
\alias{BasicFilter-class}
\alias{Cds_chromFilter}
\alias{Cds_idFilter}
\alias{Cds_nameFilter}
\alias{Cds_strandFilter}
\alias{EnsemblFilter}
\alias{GRangesFilter}
\alias{GRangesFilter-class}
\title{Functions for filter objects}
\usage{
AccnumFilter(value, condition = "==")
AliasFilter(value, condition = "==")
Cds_chromFilter(value, condition = "==")
Cds_idFilter(value, condition = "==")
Cds_nameFilter(value, condition = "==")
Cds_strandFilter(value, condition = "==")
EnsemblFilter(value, condition = "==")

GRangesFilter(value)
}
\arguments{
\item{value}{value of the filter}

\item{condition}{one of "==", "!=", "startsWith", ">", "<", ">=", "<="}
}
\description{
These functions are used to create filter objects for genomic extrators.
}
\details{
All these filters except \code{GRangesFilter} extend \code{BasicFilter}.
}
\examples{
## filter by ensembl 
EnsemblFilter("ENSG00000171862")

## filter by gene symbol start with "BRAC"
SymbolFilter("BRCA", "startsWith")

## filter by GRanges
GRangesFilter(as("chr10:87869000-87876000", "GRanges"))

## filter by transcript start position
Tx_startFilter(87863438,">")
}

