% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\docType{methods}
\name{transcripts_tbl}
\alias{transcripts_tbl}
\alias{transcripts,src_organism-method}
\alias{exons_tbl}
\alias{exons,src_organism-method}
\alias{cds_tbl}
\alias{cds,src_organism-method}
\alias{genes_tbl}
\alias{genes,src_organism-method}
\alias{promoters_tbl}
\alias{promoters,src_organism-method}
\alias{transcriptsBy_tbl}
\alias{transcriptsBy,src_organism-method}
\alias{exonsBy_tbl}
\alias{exonsBy,src_organism-method}
\alias{cdsBy_tbl}
\alias{cdsBy,src_organism-method}
\alias{intronsByTranscript_tbl}
\alias{intronsByTranscript,src_organism-method}
\alias{fiveUTRsByTranscript_tbl}
\alias{fiveUTRsByTranscript,src_organism-method}
\alias{threeUTRsByTranscript_tbl}
\alias{threeUTRsByTranscript,src_organism-method}
\title{Extract genomic features from src_organism object}
\usage{
transcripts_tbl(x, filter = NULL)

\S4method{transcripts}{src_organism}(x, filter = NULL)

exons_tbl(x, filter = NULL)

\S4method{exons}{src_organism}(x, filter = NULL)

cds_tbl(x, filter = NULL)

\S4method{cds}{src_organism}(x, filter = NULL)

genes_tbl(x, filter = NULL)

\S4method{genes}{src_organism}(x, filter = NULL)

promoters_tbl(x, upstream, downstream, filter = NULL)

\S4method{promoters}{src_organism}(x, upstream, downstream, filter = NULL)

transcriptsBy_tbl(x, by = c("gene", "exon", "cds"), filter = NULL)

\S4method{transcriptsBy}{src_organism}(x, by = c("gene", "exon", "cds"),
  filter = NULL)

exonsBy_tbl(x, by = c("tx", "gene"), filter = NULL)

\S4method{exonsBy}{src_organism}(x, by = c("tx", "gene"), filter = NULL)

cdsBy_tbl(x, by = c("tx", "gene"), filter = NULL)

\S4method{cdsBy}{src_organism}(x, by = c("tx", "gene"), filter = NULL)

intronsByTranscript_tbl(x, filter = NULL)

\S4method{intronsByTranscript}{src_organism}(x, filter = NULL)

fiveUTRsByTranscript_tbl(x, filter = NULL)

\S4method{fiveUTRsByTranscript}{src_organism}(x, filter = NULL)

threeUTRsByTranscript_tbl(x, filter = NULL)

\S4method{threeUTRsByTranscript}{src_organism}(x, filter = NULL)
}
\arguments{
\item{x}{A \code{src_organism} object}

\item{filter}{Either NULL or a named list of vectors to be used to
restrict the output. Filter can also be a \code{\link{GRanges}} object
using "GRangesFilter" (see examples).}

\item{upstream}{For \code{promoters()}: An integer(1) value indicating
the number of bases upstream from the transcription start site.}

\item{downstream}{For \code{promoters()}: An integer(1) value indicating
the number of bases downstream from the transcription start site.}

\item{by}{One of "gene", "exon", "cds" or "tx". Determines the grouping.}
}
\value{
functions with \code{_tbl} return a \code{\link[tibble]{tibble}}
    object, other methods return a \code{\link{GRanges}} or
    \code{\link{GRangesList}} object.
}
\description{
Generic functions to extract genomic features from an object. This
page documents the methods for \code{\link{src_organism}} objects
only.
}
\details{
These are the main functions for extracting transcript information
from a \code{\link{src_organism}} object, inherited from
\code{\link[GenomicFeatures]{transcripts}} in
\code{GenomicFeatures} package. Two versions of results are
provided: \code{\link[tibble]{tibble}} (\code{transcripts_tbl()})
and \code{\link{GRanges}} or \code{\link{GRangesList}}
(\code{transcripts()}).
}
\examples{
\dontrun{src <- src_ucsc("human")}
src <- src_organism(dbpath=hg38light())

## transcript coordinates with filter in tibble format
filters <- list(SymbolFilter(c("A1BG", "CDH2")))
transcripts_tbl(src, filters)

transcripts_tbl(src, list(SymbolFilter("SNORD", "startsWith")))
transcripts_tbl(src, list(GoFilter("GO:0005615")))
transcripts_tbl(src, filter=list(
     SymbolFilter("SNORD", "startsWith"),
     TxStartFilter(25070000, "<")
))
## transcript coordinates with filter in granges format
filters <- list(GRangesFilter(GenomicRanges::GRanges("chr15:1-25070000")))
transcripts(src, filters)

## promoters
promoters(src, upstream=100, downstream=50,
          filter = list(SymbolFilter("ADA")))

## transcriptsBy
transcriptsBy(src, by = "exon", filter = list(SymbolFilter("ADA")))

## exonsBy
exonsBy(src, filter = list(SymbolFilter("ADA")))

## intronsByTranscript
intronsByTranscript(src, filter = list(SymbolFilter("ADA")))

## fiveUTRsByTranscript
fiveUTRsByTranscript(src, filter = list(SymbolFilter("ADA")))

}
\seealso{
\code{\link{src_organism}} for creating a \code{src_organism}
    object.
}
\author{
Yubo Cheng.
}
