% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\docType{methods}
\name{keytypes,src_organism-method}
\alias{columns,src_organism-method}
\alias{keys,src_organism-method}
\alias{keytypes,src_organism-method}
\alias{select,src_organism-method}
\title{Using the "select" interface on src_organism objects}
\usage{
\S4method{keytypes}{src_organism}(x)

\S4method{columns}{src_organism}(x)

\S4method{keys}{src_organism}(x, keytype, ...)

\S4method{select}{src_organism}(x, keys, columns, keytype)
}
\arguments{
\item{x}{a src_organism object}

\item{keytype}{specifies the kind of keys that will be returned. By default 
keys will return the keys for schema of the \code{\link{src_organism}} 
object.}

\item{keys}{the keys to select records for from the database. All possible 
keys are returned by using the \code{keys} method.}

\item{columns}{the columns or kinds of things that can be retrieved from
the database. As with keys, all possible columns are returned by using  
the \code{columns} method.}
}
\description{
select, columns and keys can be used together to extract data for a 
\code{\link{src_organism}} object.
}
\examples{
src <- src_organism("TxDb.Hsapiens.UCSC.hg38.knownGene")
keytype <- "symbol"
keys <- c("PTEN", "BRCA1")
columns <- c("entrez", "tx_id", "tx_name","exon_id")
select(src, keys, columns, keytype)

}
\seealso{
\code{\link{AnnotationDb-class}} for more descriptsion of methods 
    \code{select}, \code{keytypes}, \code{keys} and \code{columns}.
    
    \code{\link{src_organism}} for creating a \code{src_organism} 
    object.
    
    \code{\link[Organism.dplyr]{transcripts_tbl}} for generic functions to 
    extract genomic features from a \code{src_organism} object.
}

