% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R, R/extractors.R
\docType{methods}
\name{src_organism}
\alias{select_.tbl_organism}
\alias{src_organism}
\alias{src_ucsc}
\alias{transcripts,src_organism-method}
\title{Create a dplyr view into an org package}
\usage{
src_organism(org, txdb)

src_ucsc(organism, genome = NULL, id = NULL, verbose = TRUE)

\method{select_}{tbl_organism}(.data, ...)

\S4method{transcripts}{src_organism}(x, filter)
}
\arguments{
\item{org}{character(1) naming a \code{org.*} package (e.g.,
\code{org.Hs.eg.db}) or \code{OrgDb} object instantiating the
content of an \code{org.*} pacakge.}

\item{txdb}{character(1) naming a \code{TxDb.*} package (e.g.,
\code{TxDb.Hsapiens.UCSC.hg38.knownGene}) or \code{TxDb}
object instantiating the content of a \code{TxDb.*} pacakge.}

\item{x}{object created by src_organism}

\item{filter}{Either NULL or a named list of vectors to be used to 
restrict the output.}
}
\value{
A dplyr \code{tbl_sql} instance representing the data
    table.
}
\description{
The view provides a convenient way to map between gene, transcript,
and protein identifiers.

Generic functions to extract genomic features from an object.
}
\details{
This function provides a way to map between identifiers. It does
not include details of GO evidence and ontology (see
\code{\link{tbl_go}}), and does not provide genomic coordinates
(see \code{\link{tbl_txdb}}.

This function is meant to be a building block for
\code{\link{src_organism}}, which provides an integrated
presentation of identifiers and genomic coordinates.



Create a dplyr view integrating org.* and TxDb.* information
}
\examples{
# human
library(org.Hs.eg.db)
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
organism <- src_organism(org.Hs.eg.db, TxDb.Hsapiens.UCSC.hg38.knownGene)
src_tbls(organism)
id <- tbl(organism, "id")
id
id \%>\% dplyr::select(ensembl, symbol) \%>\% filter(symbol == "PTEN")
inner_join(id, tbl(organism, "ranges_tx")) \%>\% filter(symbol == "PTEN") \%>\%
     dplyr::select(entrez, symbol, tx_id, tx_start, tx_end)


organism <- src_organism("org.Hs.eg.db", "TxDb.Hsapiens.UCSC.hg38.knownGene")
filter <- list(symbol=c("PTEN", "BRCA1"), 
               entrez="5728", 
               go=c("GO:0000079", "GO:0001933"))
transcripts(organism, filter)

}

