% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{tbl_txdb}
\alias{tbl_txdb}
\title{Create a dplyr view of genomic coordinates}
\usage{
tbl_txdb(txdb)
}
\arguments{
\item{txdb}{character(1) naming a \code{TxDb.*} package (e.g.,
\code{TxDb.Hsapiens.UCSC.hg38.knownGene}) or \code{TxDb}
object instantiating the content of a \code{TxDb.*} pacakge.}
}
\description{
Create a dplyr view of genomic coordinates
}
\examples{
txdb = "TxDb.Hsapiens.UCSC.hg38.knownGene"
tbl_txdb(txdb) \%>\% filter(entrez == "1") \%>\% 
dplyr::select(txid, txname, txstart, txend)

org = "org.Hs.eg.db"
idmap = tbl_org_idmap(org)
transcript = tbl_txdb(txdb)
inner_join(transcript, idmap, copy = TRUE) \%>\% filter(symbol == "PTEN") \%>\% 
dplyr::select(entrez, symbol, txid, txname, txstart, txend)

entrez <- idmap \%>\% filter(symbol == "PTEN") \%>\% dplyr::select(entrez, symbol)
inner_join(transcript, entrez, copy = TRUE) \%>\% filter(symbol == "PTEN") \%>\% 
dplyr::select(entrez, symbol, txid, txname, txstart, txend)

}

