% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hg38light}
\alias{hg38light}
\alias{mm10light}
\title{Utilities used in examples, vignettes, and tests}
\usage{
hg38light()

mm10light()
}
\value{
character(1) file path to the trimmed-down data base
}
\description{
These functions are primarily for illustrating
functionality. \code{hg38light()} and \code{mm10light()} provide
access to trimmed-down versions of Organism.dplyr data based
derived from the TxDb.Hsapiens.UCSC.hg38.knownGene and
TxDb.Mmusculus.UCSC.mm10.ensGene data bases.
}
\examples{
hg38light()
mm10light()
}
