% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src_organism}
\alias{src_organism}
\title{Create a dplyr view integrating org.* and TxDb.* information}
\usage{
src_organism(org, txdb)
}
\arguments{
\item{org}{character(1) naming a \code{org.*} package (e.g.,
\code{org.Hs.eg.db}) or \code{OrgDb} object instantiating the
content of an \code{org.*} pacakge.}

\item{txdb}{character(1) naming a \code{TxDb.*} package (e.g.,
\code{TxDb.Hsapiens.UCSC.hg38.knownGene}) or \code{TxDb}
object instantiating the content of a \code{TxDb.*} pacakge.}
}
\description{
Create a dplyr view integrating org.* and TxDb.* information
}

