% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\docType{methods}
\name{src_organism}
\alias{select_.tbl_organism}
\alias{seqinfo,src_organism-method}
\alias{src_organism}
\alias{src_tbls.src_organism}
\alias{src_ucsc}
\alias{supportedOrganisms}
\alias{tbl.src_organism}
\title{Create a sqlite database from an org package and a txdb package}
\usage{
src_organism(txdb = NULL, dbpath = NULL)

src_ucsc(organism, genome = NULL, id = NULL, dbpath = NULL,
  verbose = TRUE)

supportedOrganisms()

\method{select_}{tbl_organism}(.data, ...)

\method{src_tbls}{src_organism}(x)

\method{tbl}{src_organism}(src, ...)

\S4method{seqinfo}{src_organism}(x)
}
\arguments{
\item{txdb}{character(1) naming a \code{TxDb.*} package (e.g.,
\code{TxDb.Hsapiens.UCSC.hg38.knownGene}) or \code{TxDb}
object instantiating the content of a \code{TxDb.*} pacakge.}

\item{dbpath}{path and file name where SQLite file will be accessed
or created if not already exists.}

\item{organism}{organism or common name}

\item{genome}{genome name}

\item{id}{choose from "knownGene", "ensGene" and "refGene"}

\item{verbose}{logical. Should R report extra information on progress?
Default is TRUE.}

\item{.data}{A tbl.}

\item{...}{Comma separated list of unquoted expressions. You can treat
variable names like they are positions. Use positive values to select
variables; use negative values to drop variables.}

\item{x}{A src_organism object}

\item{src}{A src_organism object}
}
\value{
\code{src_organism()} and \code{src_ucsc()} returns a dplyr 
    \code{src_sqlite} instance representing the data tables.
}
\description{
The database provides a convenient way to map between gene, transcript,
and protein identifiers.
}
\details{
\code{src_organism()} and \code{src_ucsc()} are meant to be a building block 
for \code{\link{src_organism}}, which provides an integrated
presentation of identifiers and genomic coordinates.

\code{src_organism()} creates a dplyr database integrating org.* and TxDb.* 
information by given TxDb. And \code{src_ucsc()} creates the database by 
given organism name, genome and/or id.

supportedOrganisms() provides all supported organisms in this package with 
corresponding OrgDb and TxDb.
}
\examples{
## create human sqlite database with TxDb.Hsapiens.UCSC.hg38.knownGene and
## corresponding org.Hs.eg.db
src <- src_organism("TxDb.Hsapiens.UCSC.hg38.knownGene")

## query using dplyr
inner_join(tbl(src, "id"), tbl(src, "id_go")) \%>\% 
     filter(symbol == "PTEN") \%>\% 
     dplyr::select(entrez, ensembl, symbol, go, evidence, ontology)

## create human sqlite database using hg38 genome
human <- src_ucsc("human")

## all supported organisms with corresponding OrgDb and TxDb
supportedOrganisms()

## Look at all available tables
src_tbls(src)

## Look at data in table "id"
tbl(src, "id")

## Look at fields of one table 
colnames(tbl(src, "id"))

## seqinfo of src_organism object
seqinfo(src)

}
\seealso{
\code{\link{dplyr}} for details about using \code{dplyr} to 
    manipulate data. 
    
    \code{\link[Organism.dplyr]{transcripts_tbl}} for generic functions to 
    extract genomic features from a \code{src_organism} object.
    
    \code{\link[Organism.dplyr]{select,src_organism-method}} for "select" 
    interface on \code{src_organism} objects.
}

