% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src_organism}
\alias{select_.tbl_organism}
\alias{src_organism}
\title{Create a dplyr view into an org package}
\usage{
src_organism(org, txdb)

\method{select_}{tbl_organism}(.data, ...)
}
\arguments{
\item{org}{character(1) naming a \code{org.*} package (e.g.,
\code{org.Hs.eg.db}) or \code{OrgDb} object instantiating the
content of an \code{org.*} pacakge.}

\item{txdb}{character(1) naming a \code{TxDb.*} package (e.g.,
\code{TxDb.Hsapiens.UCSC.hg38.knownGene}) or \code{TxDb}
object instantiating the content of a \code{TxDb.*} pacakge.}
}
\value{
A dplyr \code{tbl_sql} instance representing the data
    table.
}
\description{
The view provides a convenient way to map between gene, transcript,
and protein identifiers.
}
\details{
This function provides a way to map between identifiers. It does
not include details of GO evidence and ontology (see
\code{\link{tbl_go}}), and does not provide genomic coordinates
(see \code{\link{tbl_txdb}}.

This function is meant to be a building block for
\code{\link{src_organism}}, which provides an integrated
presentation of identifiers and genomic coordinates.



Create a dplyr view integrating org.* and TxDb.* information
}
\examples{
organism <- src_organism("org.Hs.eg.db", "TxDb.Hsapiens.UCSC.hg38.knownGene")
src_tbls(organism)
tbl <- tbl(organism, "id")
tbl
tbl \%>\% dplyr::select(ensembl, symbol) \%>\% filter(symbol == "PTEN")
inner_join(tbl, tbl(organism, "txdb")) \%>\% filter(symbol == "PTEN") \%>\%
     dplyr::select(entrez, symbol, txid, txname, txstart, txend)

}

