% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src_organism}
\alias{select_.tbl_organism}
\alias{src_by_organism}
\alias{src_organism}
\title{Create a dplyr view into an org package}
\usage{
src_organism(org, txdb)

src_by_organism(organism)

\method{select_}{tbl_organism}(.data, ...)
}
\arguments{
\item{org}{character(1) naming a \code{org.*} package (e.g.,
\code{org.Hs.eg.db}) or \code{OrgDb} object instantiating the
content of an \code{org.*} pacakge.}

\item{txdb}{character(1) naming a \code{TxDb.*} package (e.g.,
\code{TxDb.Hsapiens.UCSC.hg38.knownGene}) or \code{TxDb}
object instantiating the content of a \code{TxDb.*} pacakge.}
}
\value{
A dplyr \code{tbl_sql} instance representing the data
    table.
}
\description{
The view provides a convenient way to map between gene, transcript,
and protein identifiers.
}
\details{
This function provides a way to map between identifiers. It does
not include details of GO evidence and ontology (see
\code{\link{tbl_go}}), and does not provide genomic coordinates
(see \code{\link{tbl_txdb}}.

This function is meant to be a building block for
\code{\link{src_organism}}, which provides an integrated
presentation of identifiers and genomic coordinates.



Create a dplyr view integrating org.* and TxDb.* information
}
\examples{
# human
library(org.Hs.eg.db)
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
organism <- src_organism(org.Hs.eg.db, TxDb.Hsapiens.UCSC.hg38.knownGene)
src_tbls(organism)
id <- tbl(organism, "id")
id
id \%>\% dplyr::select(ensembl, symbol) \%>\% filter(symbol == "PTEN")
inner_join(id, tbl(organism, "ranges_tx")) \%>\% filter(symbol == "PTEN") \%>\%
     dplyr::select(entrez, symbol, txid, start, end)

# mouse
organism <- src_organism("org.Mm.eg.db", "TxDb.Mmusculus.UCSC.mm10.knownGene")
inner_join(tbl(organism, "id"), tbl(organism, "ranges_gene")) \%>\% 
     filter(entrez == "11287") \%>\%
     dplyr::select(entrez, symbol, chrom, start, end)

# rat
organism <- src_organism("org.Rn.eg.db", "TxDb.Rnorvegicus.UCSC.rn4.ensGene")
inner_join(tbl(organism, "id"), tbl(organism, "ranges_gene"), by = c("ensembl" = "geneid")) \%>\% 
     filter(entrez.y == "ENSRNOG00000028896") \%>\%
     dplyr::select(entrez.x, symbol, chrom, start, end)

# transcripts()
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene
transcripts(txdb, filter=list(tx_name="uc001aaa.3"))

human <- src_organism("org.Hs.eg.db", "TxDb.Hsapiens.UCSC.hg38.knownGene")

tx <- inner_join(tbl(human, "ranges_tx"), tbl(human, "id_ranges")) \%>\%
    dplyr::select(chrom, start, end, strand, txid, txname)

tx \%>\% filter(txname == "uc001aaa.3")


}

