% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R, R/extractors.R
\docType{methods}
\name{src_organism}
\alias{cds,src_organism-method}
\alias{cdsBy,src_organism-method}
\alias{cds_tbl}
\alias{exons,src_organism-method}
\alias{exonsBy,src_organism-method}
\alias{exons_tbl}
\alias{fiveUTRsByTranscript,src_organism-method}
\alias{genes,src_organism-method}
\alias{genes_tbl}
\alias{intronsByTranscript,src_organism-method}
\alias{promoters,src_organism-method}
\alias{promoters_tbl}
\alias{select_.tbl_organism}
\alias{src_organism}
\alias{src_ucsc}
\alias{threeUTRsByTranscript,src_organism-method}
\alias{transcripts,src_organism-method}
\alias{transcriptsBy,src_organism-method}
\alias{transcriptsBy_tbl}
\alias{transcripts_tbl}
\title{Create a dplyr view into an org package}
\usage{
src_organism(org = NULL, txdb = NULL, dbpath = NULL)

src_ucsc(organism, genome = NULL, id = NULL, dbpath = NULL,
  verbose = TRUE)

\method{select_}{tbl_organism}(.data, ...)

transcripts_tbl(x, filter = NULL)

\S4method{transcripts}{src_organism}(x, filter = NULL)

exons_tbl(x, filter = NULL)

\S4method{exons}{src_organism}(x, filter = NULL)

cds_tbl(x, filter = NULL)

\S4method{cds}{src_organism}(x, filter = NULL)

genes_tbl(x, filter = NULL)

\S4method{genes}{src_organism}(x, filter = NULL)

promoters_tbl(x, upstream, downstream, filter = NULL)

\S4method{promoters}{src_organism}(x, upstream, downstream, filter = NULL)

transcriptsBy_tbl(x, by = c("gene", "exon", "cds"), filter = NULL, schema)

\S4method{transcriptsBy}{src_organism}(x, by = c("gene", "exon", "cds"),
  filter = NULL)

\S4method{exonsBy}{src_organism}(x, by = c("tx", "gene"), filter = NULL)

\S4method{cdsBy}{src_organism}(x, by = c("tx", "gene"), filter = NULL)

\S4method{intronsByTranscript}{src_organism}(x, filter = NULL)

\S4method{fiveUTRsByTranscript}{src_organism}(x, filter = NULL)

\S4method{threeUTRsByTranscript}{src_organism}(x, filter = NULL)
}
\arguments{
\item{org}{character(1) naming a \code{org.*} package (e.g.,
\code{org.Hs.eg.db}) or \code{OrgDb} object instantiating the
content of an \code{org.*} pacakge.}

\item{txdb}{character(1) naming a \code{TxDb.*} package (e.g.,
\code{TxDb.Hsapiens.UCSC.hg38.knownGene}) or \code{TxDb}
object instantiating the content of a \code{TxDb.*} pacakge.}

\item{organism}{}

\item{genome}{}

\item{id}{}

\item{x}{A src_organism object}

\item{filter}{Either NULL or a named list of vectors to be used to
restrict the output.}

\item{upstream}{For \code{promoters()}: An integer(1) value indicating 
the number of bases upstream from the transcription start site.}

\item{downstream}{For \code{promoters()}: An integer(1) value indicating
the number of bases downstream from the transcription start site.}

\item{by}{One of "gene", "exon", "cds" or "tx". Determines the grouping.}
}
\value{
A dplyr \code{tbl_sql} instance representing the data
    table.
}
\description{
The view provides a convenient way to map between gene, transcript,
and protein identifiers.

Generic functions to extract genomic features from an object.
}
\details{
This function provides a way to map between identifiers. It does
not include details of GO evidence and ontology (see
\code{\link{tbl_go}}), and does not provide genomic coordinates
(see \code{\link{tbl_txdb}}.

This function is meant to be a building block for
\code{\link{src_organism}}, which provides an integrated
presentation of identifiers and genomic coordinates.



Create a dplyr view integrating org.* and TxDb.* information
}
\examples{
# human
library(org.Hs.eg.db)
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
organism <- src_organism(org.Hs.eg.db, TxDb.Hsapiens.UCSC.hg38.knownGene)
src_tbls(organism)
id <- tbl(organism, "id")
id
id \%>\% dplyr::select(ensembl, symbol) \%>\% filter(symbol == "PTEN")
inner_join(id, tbl(organism, "ranges_tx")) \%>\% filter(symbol == "PTEN") \%>\%
     dplyr::select(entrez, symbol, tx_id, tx_start, tx_end)


human <- src_ucsc("human")

organism <- src_ucsc("human")
filters <- list(symbol=c("PTEN", "BRCA1"),
               entrez="5728",
               go=c("GO:0000079", "GO:0001933"))
transcripts(organism, filters)

exons(organism, filter=list(symbol="PTEN"))

promoters(organism, upstream=100, downstream=50,
          filter = list(symbol="BRCA1"))

transcriptsBy(organism, by = "gene", filter = list(symbol="PTEN"))

exonsBy(organism, by = "gene", filter = list(symbol="PTEN"))

cdsBy(organism, by = "gene", filter = list(symbol="PTEN"))

intronsByTranscript(organism, filter = list(symbol="PTEN"))

fiveUTRsByTranscript(organism, filter = list(symbol="PTEN"))

threeUTRsByTranscript(organism, filter = list(symbol="PTEN"))

}

