% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{tbl_org}
\alias{tbl_org}
\title{Create a dplyr view into an org package}
\usage{
tbl_org(org)
}
\arguments{
\item{org}{character(1) naming a \code{org.*} package (e.g.,
\code{org.Hs.eg.db}) or \code{OrgDb} object instantiating the
content of an \code{org.*} pacakge.}
}
\value{
A dplyr \code{tbl_sql} instance representing the data
    table.
}
\description{
The view provides a convenient way to map between gene, transcript,
and protein identifiers.
}
\details{
This function provides a way to map between identifiers. It does
not include details of GO evidence and ontology (see
\code{\link{tbl_go}}), and does not provide genomic coordinates
(see \code{\link{tbl_txdb}}.

This function is meant to be a building block for
\code{\link{src_organism}}, which provides an integrated
presentation of identifiers and genomic coordinates.
}
\examples{
tbl = tbl_org("org.Hs.eg.db")
tbl
tbl \%>\% select(ensembl, symbol) \%>\% filter(symbol == "PTEN")
symbols <- paste0("BRCA", 1:2)
tbl \%>\% select(ensembl, symbol) \%>\% filter(symbol \%in\% symbols)

}

