% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{tbl_txdb}
\alias{tbl_txdb}
\title{Create a dplyr view of genomic coordinates}
\usage{
tbl_txdb(txdb)
}
\arguments{
\item{txdb}{character(1) naming a \code{TxDb.*} package (e.g.,
\code{TxDb.Hsapiens.UCSC.hg38.knownGene}) or \code{TxDb}
object instantiating the content of a \code{TxDb.*} pacakge.}
}
\description{
Create a dplyr view of genomic coordinates
}

